Program 1 Part 4
Write a C Program To Read Two Matrices A (m x n) and B (p x q) and Computes the Product A and B. Read Matrix A and Matrix B in Row Major Order Respectively. Print Both The Input Matrices and Resultant Matrix With Suitable Headings and Output Should Be in Matrix Format Only. Program Must Check The Compatibility of Orders of The Matrices for Multiplication. Report Appropriate Message in Case of Incompatibility.